Adjust api controller structure
